% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_had.R
\name{did_had}
\alias{did_had}
\title{Outer function for DIDHetAdoption package}
\usage{
did_had(
  df,
  outcome,
  group,
  time,
  treatment,
  effects = 1,
  placebo = 0,
  level = 5,
  kernel = "uni",
  graph_off = FALSE
)
}
\arguments{
\item{df}{A data.frame object}

\item{outcome}{Outcome variable}

\item{group}{Group Variable}

\item{time}{Time variable}

\item{treatment}{Treatment variable}

\item{effects}{effects}

\item{placebo}{placebo}

\item{level}{level}

\item{kernel}{kernel}

\item{graph_off}{graph_off}
}
\description{
Estimation of DiD estimators in Heterogeneous adoption designs (de Chaisemartin and D'Haultfoeuille, 2024).
}
\examples{
A <- 0; print(A)
}
